Robot code upgraded
I have changed the scan number to 12, this in theory should make it
better BUT at the moment is slower to compensate I have jigged the SD
and probs around but this has had a negative effect on the accuracy. The
SD and probs etc ideally these would be a function of the number of
particles. (Just needs adjusting until a good solution is found). I have
added lots of notes to robot/realRobot... and a few to localise